+++ backend/Erabikata.Backend/Models/Output/WordRank.cs
@@ -2,4 +2,4 @@ namespace Erabikata.Backend.Models.Output;
 
 public record WordRank(int Id, long? Rank);
 
-public record WordRankInfo(int Id, int Rank, uint Count, string Text);
+public record WordRankInfo(int Id, int Rank, ulong Count, string Text);
